feat: add shared signed URL cache hook
